+++ Trainer_v5/Trainer.Source/SDK/DefaultStyles.cs
@@ -4,21 +4,23 @@ namespace Trainer_v5.Trainer.Source.SDK
 {
 	public static class ComponentStyleHelper
 	{
-		private static readonly ComponentStyle DefaultStyle = new ComponentStyle(32);
+		private static readonly ComponentStyle _defaultStyle = new ComponentStyle(32);
 
 		public static ComponentStyle GetStyle(this Component self)
 		{
-			return DefaultStyle;
+			return _defaultStyle;
 		}
 	}
 
 	public class ComponentStyle
 	{
-		public readonly int DefaultHeight;
+		private readonly int _defaultHeight;
+		
+		public int DefaultHeight => _defaultHeight;
 
 		public ComponentStyle(int defaultHeight)
 		{
-			DefaultHeight = defaultHeight;
+			_defaultHeight = defaultHeight;
 		}
 	}
 }